rearchitect code
