+++ tools/backend/src/app.ts
@@ -15,15 +15,15 @@ import { staticFrontendDirRelative } from './utils.js';
 import { cors } from 'hono/cors';
 import { serveStatic } from '@hono/node-server/serve-static';
 import { attachCommandManager } from './middleware/commandManager.js';
-import cardsRouter from './routes/cards.js';
-import fieldTypesRouter from './routes/fieldTypes.js';
-import linkTypesRouter from './routes/linkTypes.js';
-import templatesRouter from './routes/templates.js';
-import treeRouter from './routes/tree.js';
+import cardsRouter from './domain/cards/index.js';
+import fieldTypesRouter from './domain/fieldTypes/index.js';
+import linkTypesRouter from './domain/linkTypes/index.js';
+import templatesRouter from './domain/templates/index.js';
+import treeRouter from './domain/tree/index.js';
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { isSSGContext } from './export.js';
-import resourcesRouter from './routes/resources.js';
+import resourcesRouter from './domain/resources/index.js';
 
 /**
  * Create the Hono app for the backend
