+++ tools/backend/src/domain/fieldTypes/index.ts
@@ -1,0 +1,49 @@
+/**
+  Cyberismo
+  Copyright © Cyberismo Ltd and contributors 2025
+  This program is free software: you can redistribute it and/or modify it under
+  the terms of the GNU Affero General Public License version 3 as published by
+  the Free Software Foundation.
+  This program is distributed in the hope that it will be useful, but WITHOUT
+  ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+  FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+  details. You should have received a copy of the GNU Affero General Public
+  License along with this program. If not, see <https://www.gnu.org/licenses/>.
+*/
+
+import { Hono } from 'hono';
+import * as fieldTypeService from './service.js';
+
+const router = new Hono();
+
+/**
+ * @swagger
+ * /api/fieldTypes:
+ *   get:
+ *     summary: Returns a list of all field types in the defined project.
+ *     description: List of field types includes all field types in the project with all their details
+ *     responses:
+ *       200:
+ *        description: Object containing the project field types. See definitions.ts/FieldTypes for the structure.
+ *       400:
+ *         description: Error in reading project details.
+ *       500:
+ *         description: project_path not set or other internal error
+ */
+router.get('/', async (c) => {
+  const commands = c.get('commands');
+
+  try {
+    const fieldTypes = await fieldTypeService.getFieldTypes(commands);
+    return c.json(fieldTypes);
+  } catch (error) {
+    return c.json(
+      {
+        error: `${error instanceof Error ? error.message : 'Unknown error'} from path ${c.get('projectPath')}`,
+      },
+      500,
+    );
+  }
+});
+
+export default router;
